backend/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
requests
google-generativeai
psycopg2-binary
python-decouple  # Optional: better env management